Remove all code except individual technique schema and validation
