--- mitiq/schema/experiments/ddd_experiments/ddd_experiment.json
@@ -1,3 +1,0 @@
-{
-    "rule": "xx"
-}
