--- mitiq/schema/experiments/ddd_experiments/ddd_experiment_with_technique.json
@@ -1,4 +1,0 @@
-{
-    "technique": "ddd",
-    "rule": "xx"
-}
