--- mitiq/schema/experiments/ddd_experiments/ddd_experiment_xyxy.json
@@ -1,3 +1,0 @@
-{
-    "rule": "xyxy"
-}
