--- mitiq/schema/experiments/ddd_experiments/ddd_experiment_yy.json
@@ -1,3 +1,0 @@
-{
-    "rule": "yy"
-}
