--- mitiq/schema/experiments/pec_experiments/pec_experiment.json
@@ -1,3 +1,0 @@
-{
-    "rule": "xx"
-}
