--- mitiq/schema/experiments/rem_experiments/rem_experiment.json
@@ -1,3 +1,0 @@
-{
-    "rule": "xx"
-}
