--- mitiq/schema/experiments/zne_experiments/zne_experiment_exp.json
@@ -1,5 +1,0 @@
-{
-    "noise_scaling_factors": [2, 4, 6, 8],
-    "noise_scaling_method": "global",
-    "extrapolation": "exponential"
-    }
