--- mitiq/schema/experiments/zne_experiments/zne_experiment_with_technique.json
@@ -1,7 +1,0 @@
-{
-    "technique": "zne",
-    "noise_scaling_factors": [2, 4, 6, 8],
-    "noise_scaling_method": "global",
-    "extrapolation": "richardson",
-    "scale_factor": 2
-    }
