--- mitiq/schema/schema/composition_schema_ordered.json
@@ -1,22 +1,0 @@
-{
-    "$schema": "http://json-schema.org/draft-07/schema#",
-    "$id": "file:////Users/ellac/Documents/Unitary_Fund/qem_schema/schema/composition_schema.json",
-    "type": "object",
-    "properties": {
-      "techniques": {
-        "type": "array",
-        "description": "Array of techniques to validate",
-        "items": {
-          "oneOf": [
-            { "$ref": "./zne_schema.json" },
-            { "$ref": "./ddd_schema.json" },
-            { "$ref": "./pec_schema.json" }
-          ]
-        },
-        "minItems": 1,
-        "uniqueItems": true
-      }
-    },
-    "required": ["techniques"],
-    "additionalProperties": false
-  }
