+++ mitiq/schema/technique_schema/ddd_schema.json
